module9-task2: refactoring
