+++ src/components/app/app.tsx
@@ -7,7 +7,7 @@ import MyList from '../../pages/my-list/my-list';
 import PageNotFound from '../../pages/page-not-found/page-not-found';
 import AddReview from '../../pages/add-review/add-review';
 import SignIn from '../../pages/login/login';
-import MoviePage from '../../pages/movie-page/movie-page';
+import FilmPage from '../../pages/film-page/film-page.tsx';
 import Player from '../../pages/player/player';
 import {AppRoute} from '../../enums/app-route';
 import {getAuthStatus} from '../../store/user-process/user-process.selectors.ts';
@@ -38,7 +38,7 @@ function App(): React.JSX.Element {
             }
           />
           <Route path={AppRoute.Films}>
-            <Route path=":id" element={<MoviePage />}/>
+            <Route path=":id" element={<FilmPage />}/>
             <Route
               path={`:id${AppRoute.Review}`}
               element={
