+++ src/components/catalog/catalog.tsx
@@ -1,9 +1,9 @@
 import GenresList from '../genre-list/genre-list';
 import FilmsList from '../films-list/films-list';
-import React, { useCallback, useState } from 'react';
-import { DEFAULT_FILM_LIST_LENGTH } from '../../constants/film-list.ts';
-import { useAppSelector } from '../../hooks/store.ts';
-import { getFilmsByGenreLength} from '../../store/films-process/films-process.selectors.ts';
+import React, {useCallback, useState} from 'react';
+import {DEFAULT_FILM_LIST_LENGTH} from '../../constants/film-list';
+import {useAppSelector} from '../../hooks/store';
+import {getFilmsByGenreLength} from '../../store/films-process/films-process.selectors';
 
 function Catalog(): React.JSX.Element {
   const stateGenreFilmsLength = useAppSelector(getFilmsByGenreLength);
