+++ src/components/tabs/tabs.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { IFilmPromoInfo } from '../../types/film-types.ts';
-import Overview from '../overview/overview.tsx';
+import { IFilmPromoInfo } from '../../types/film-types';
+import Overview from '../overview/overview';
 import Details from '../film-details/film-details';
 import Reviews from '../film-reviews/film-reviews';
-import { IReview } from '../../types/review-types.ts';
+import { IReview } from '../../types/review-types';
 
 const TABS = ['Overview', 'Details', 'Reviews'] as const;
