+++ src/constants/time.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_AUTOPLAY_TIME = 1000;
+export const SEC_PER_HOUR = 3600;
+export const SEC_PER_MINUTE = 60;
+export const MIN_PER_HOUR = 60;
