--- src/mocks/reviews.ts
@@ -1,54 +1,0 @@
-import { IReview } from '../types/review-types';
-
-export const reviewsInfo: IReview[] = [
-  {
-    id: 'ac8adb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2023-05-25T12:00:00.000Z',
-    user: 'Kate Muir',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 8
-  },
-  {
-    id: 'akfkgdb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2022-10-25T12:00:00.000Z',
-    user: 'Alexey',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 9
-  },
-  {
-    id: 'ac8adb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2023-01-02T12:00:00.000Z',
-    user: 'Lilia',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 6.6
-  },
-  {
-    id: 'ac8adb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2023-03-05T12:00:00.000Z',
-    user: 'John',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 8.9
-  },
-  {
-    id: 'ac8adb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2023-08-09T12:00:00.000Z',
-    user: 'Kate',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 4
-  },
-  {
-    id: 'ac8adb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2023-10-19T12:00:00.000Z',
-    user: 'Sergey',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 5
-  },
-  {
-    id: 'ac8adb38-0cfa-4585-a767-d3afcc557a05',
-    date: '2023-03-17T12:00:00.000Z',
-    user: 'Ivan',
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    rating: 7
-  },
-
-];
