+++ src/store/action.ts
@@ -1,4 +1,4 @@
 import { createAction } from '@reduxjs/toolkit';
-import {AppRoute} from '../enums/app-route.ts';
+import {AppRoute} from '../enums/app-route';
 
 export const redirectToRoute = createAction<AppRoute | string>('redirectToRoute');
