+++ src/store/root-reducer.ts
@@ -1,6 +1,6 @@
 import {combineReducers} from '@reduxjs/toolkit';
 import {userProcessSlice} from './user-process/user-process.slice.ts';
-import {NameSpace} from '../constants/namespaces.ts';
+import {NameSpace} from '../enums/namespace.ts';
 import {filmProcessSlice} from './film-process/film-process.slice.ts';
 import {filmsProcessSlice} from './films-process/films-process.slice.ts';
