+++ src/store/user-process/user-process.slice.test.ts
@@ -1,6 +1,6 @@
-import {AuthStatus} from '../../enums/auth-status.ts';
-import {createUser} from '../../mocks/mocks.ts';
-import {UserProcessState} from '../../types/state.ts';
+import {AuthStatus} from '../../enums/auth-status';
+import {createUser} from '../../utils/mocks';
+import {UserProcessState} from '../../types/state';
 import { checkAuthStatusAction, loginAction, logoutAction } from '../api-actions';
 import { userProcessSlice } from './user-process.slice';
