+++ src/store/user-process/user-process.slice.ts
@@ -1,9 +1,9 @@
 import {createSlice} from '@reduxjs/toolkit';
-import {NameSpace} from '../../constants/namespaces.ts';
-import {UserProcessState} from '../../types/state.ts';
-import {AuthStatus} from '../../enums/auth-status.ts';
-import {checkAuthStatusAction, loginAction, logoutAction} from '../api-actions.ts';
-import {removeToken, setToken} from '../../services/token.ts';
+import {NameSpace} from '../../enums/namespace';
+import {UserProcessState} from '../../types/state';
+import {AuthStatus} from '../../enums/auth-status';
+import {checkAuthStatusAction, loginAction, logoutAction} from '../api-actions';
+import {removeToken, setToken} from '../../services/token';
 
 const initialState: UserProcessState = {
   authorizationStatus: AuthStatus.Unknown,
@@ -13,7 +13,11 @@ const initialState: UserProcessState = {
 export const userProcessSlice = createSlice({
   name: NameSpace.User,
   initialState,
-  reducers: {},
+  reducers: {
+    setAuthStatus: (state, action) => {
+      state.authorizationStatus = action.payload as AuthStatus;
+    }
+  },
   extraReducers: (builder) => {
     builder
       .addCase(logoutAction.fulfilled, (state) => {
@@ -27,14 +31,13 @@ export const userProcessSlice = createSlice({
         state.authorizationStatus = AuthStatus.Auth;
       })
       .addCase(checkAuthStatusAction.fulfilled, (state, action) => {
-        setToken(action.payload.token);
         state.user = action.payload;
-        state.authorizationStatus = AuthStatus.Auth;
       })
       .addCase(checkAuthStatusAction.rejected, (state) => {
         removeToken();
         state.user = null;
-        state.authorizationStatus = AuthStatus.NoAuth;
       });
   }
 });
+
+export const {setAuthStatus} = userProcessSlice.actions;
